requirements.txt
Django==5.0.6
whitenoise==6.6.0
python-dotenv==1.0.1
gunicorn==22.0.0
python-telegram-bot==21.6